Enable DLL and exe builds.
